Keyboard fix, verification account done
Keyboard wont push nav bar,
verification account also done
