+++ main/bundle_of_joy/lib/auth/auth.dart
@@ -2,7 +2,7 @@ import "package:firebase_core/firebase_core.dart";
 import "package:firebase_auth/firebase_auth.dart";
 import "package:google_sign_in/google_sign_in.dart";
 import "package:flutter_facebook_login/flutter_facebook_login.dart";
-import 'package:bundle_of_joy/mother.dart';
+import 'file:///C:/Users/User/Desktop/Code/Bundle-Of-Joy/main/bundle_of_joy/lib/motherProfile/mother.dart';
 
 final FirebaseAuth _auth = FirebaseAuth.instance;
 final GoogleSignIn googleSignIn = GoogleSignIn();
