+++ main/bundle_of_joy/pubspec.yaml
@@ -48,6 +48,7 @@ dependencies:
   flutter_form_builder: ^3.14.0-alpha.4
   date_format: ^1.0.9
   fl_chart: ^0.12.0
+  pinput: ^0.2.6
 
 dev_dependencies:
   flutter_test:
